Renaming 'network' to 'session_network' because xcode can't debug
