+++ src/onionreq/builder.cpp
@@ -22,7 +22,7 @@
 #include <oxen/quic/address.hpp>
 
 #include "session/export.h"
-#include "session/network.hpp"
+#include "session/session_network.hpp"
 #include "session/onionreq/builder.h"
 #include "session/onionreq/hop_encryption.hpp"
 #include "session/onionreq/key_types.hpp"
