+++ src/session_network.cpp
@@ -1,4 +1,4 @@
-#include "session/network.hpp"
+#include "session/session_network.hpp"
 
 #include <fmt/ranges.h>
 #include <oxenc/base64.h>
@@ -24,7 +24,7 @@
 #include "session/ed25519.hpp"
 #include "session/export.h"
 #include "session/file.hpp"
-#include "session/network.h"
+#include "session/session_network.h"
 #include "session/onionreq/builder.h"
 #include "session/onionreq/builder.hpp"
 #include "session/onionreq/key_types.hpp"
