+++ tests/CMakeLists.txt
@@ -30,7 +30,7 @@ set(LIB_SESSION_UTESTS_SOURCES
 )
 
 if (ENABLE_ONIONREQ)
-    list(APPEND LIB_SESSION_UTESTS_SOURCES test_network.cpp)
+    list(APPEND LIB_SESSION_UTESTS_SOURCES test_session_network.cpp)
     list(APPEND LIB_SESSION_UTESTS_SOURCES test_onionreq.cpp)
 endif()
