+++ tests/test_onionreq.cpp
@@ -1,5 +1,5 @@
 #include <catch2/catch_test_macros.hpp>
-#include <session/network.hpp>
+#include <session/session_network.hpp>
 #include <session/onionreq/hop_encryption.hpp>
 #include <session/onionreq/parser.hpp>
